construction and inf-edge-conf
inf-edge-conf
when t is segment and sgn is negative,
we have tentatively return true for the case:
q or r or s is a segment and sgn is negative.
Constructions:
For the new bisector type,
now we have bisector lines, for the case:
p and q both segments sharing an end-point.