different parabola arcs for pqrt, case p is point
This is a bug fix for the following input br80.cin.

s       10      120     60      20
s       60      40      70      60
s       30      110     100     40

A similar fix is probably required for fecf pqrt case,
when p is segment and q is a point. The fix works when
vertices (p,q,r) and (q,p,inf) are not on different arcs
of the Linf parabola of p and q.

Explanation for arcs:
The Linf-perpendicular line to q that passes through p partitions
the Linf-parabola of p and q in two arcs.

Signed-off-by: Panagiotis Cheilaris <philaris@cs.ntua.gr>