benchmarks for Linf and L2 adapted from Menelaos'
The benchmarks do not support crossing segmetns.

The input should be fed from stdin and should
be in the the format of "p x y" or "s x1 y1 x2 y2"
per line. Do not give the total number of points
and segments in the first line (this is in contrast
with Menelaos' benchmark expected input).

Example input:
p 20 34
s 10 34 78 1
s 23 18 91 22

Example run:
n=600000 ;
 rbox $n D2 s B180 | tail -n $n | awk '{print "p " $1 " " $2 }' |
 time benchmark-linf

Signed-off-by: Panagiotis Cheilaris <philaris@cs.ntua.gr>