ipescript ipe2plg converts ipe to plg format
The script ipe2plg converts all paths and points
in the first page of the ipe file given as the command
line argument to the polygon .plg format of CGAL.
The result is printed to the standard output.
Use redirection to save to a .plg file.

The file ipe2plg.lua has to be put in the appropriate
place, e.g., ~/.ipe/scripts/ and here is an example use:

ipescript ipe2plg test1.ipe > test1.plg

Signed-off-by: Panagiotis Cheilaris <philaris@cs.ntua.gr>