support line optimization in first intersection
Addition of is_line_optimization boolean member variable in object
of class Polychainline_2. Suppose this variable of an obect is set
to true. If this object is tried for an intersection with another
Polychainline object pcl, using the function
first_intersection_point_with, then instead the function
line_first_intersection_point_with is used.

Signed-off-by: Panagiotis Cheilaris <philaris@cs.ntua.gr>