fix p,q,inf vertex conflict when p, q share coord
This fixes a problem when a floating point filtering kernel is
used and the input consists of non-integer numbers. The problem
manifestated with input n2scaled.cin:

s 4.703 61.000  4.700 60.996
s 4.732 60.997  4.735 61.000

(s A B, s C D)

The vertex conflict predicate (A, D, inf ; AB) returned NEGATIVE.
This was because an oriented side predicate for line AD and point
A of segment AB was computed. This predicate does not have to be
run, because we already know that A is an endpoint of AB.

Signed-off-by: Panagiotis Cheilaris <philaris@cs.ntua.gr>